requirements.txt
spacy==3.7.2
transformers==4.36.2
torch==2.2.1
python-docx==1.0.1
PyMuPDF==1.23.8
scikit-learn==1.3.2
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
mistralai
python-dotenv
docx2pdf==0.1.8
reportlab==4.0.7
# spacy-udpipe
# stanza
# sentence_tokenize